fix(remotecontrol): widen instance IDs to long for Unity 6.5 EntityId / fix(vrchatavatartransfer): resolve asset identity the same way / chore(livestudio): refresh Studio Template dependencies
Unity 6.3 (6000.3) replaced GetInstanceID / Resources.InstanceIDToObject with
their EntityId counterparts, and Unity 6.5 (6000.5) promoted both the old API
and the EntityId<->int implicit conversions to Obsolete errors. Callers now go
through ExposedObjectUtility and carry the ID as long, so the version split
stays in one place and survives EntityId outgrowing 32 bits.

+++ Packages/RemoteControl/Runtime/Handlers/ExposedObjectHandler.cs
@@ -917,7 +917,7 @@ namespace Lilium.RemoteControl
 
             // 最終フォールバック: 数値 instanceId として Unity の内部 API から逆引きして
             // 未登録の UnityEngine.Object を一時的な ExposedObjectHandle にラップする（レジストリ登録しない）
-            if (int.TryParse(id, out var unityInstanceId))
+            if (long.TryParse(id, out var unityInstanceId))
             {
                 var unityObj = ExposedObjectUtility.InstanceIDToObject(unityInstanceId);
                 if (unityObj != null)
